Rename `spring-ai-autoconfigure-model-chat-memory-redis` to `spring-ai-autoconfigure-model-chat-memory-repository-redis`
