+++ auto-configurations/models/chat/memory/repository/spring-ai-autoconfigure-model-chat-memory-repository-redis/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-org.springframework.ai.model.chat.memory.autoconfigure.ChatMemoryAutoConfiguration
+org.springframework.ai.model.chat.memory.redis.repository.autoconfigure.RedisChatMemoryAutoConfiguration
