+++ starters/spring-ai-starter-model-chat-memory-repository-redis/pom.xml
@@ -44,7 +44,7 @@
 
 		<dependency>
 			<groupId>org.springframework.ai</groupId>
-			<artifactId>spring-ai-autoconfigure-model-chat-memory-redis</artifactId>
+			<artifactId>spring-ai-autoconfigure-model-chat-memory-repository-redis</artifactId>
 			<version>${project.parent.version}</version>
 		</dependency>
